lint fix - update package name
